Refactor table refresh tests into category traits following Spark DSv2 pattern
Restructure the monolithic DeltaTableRefreshAndPinningSuite (1686 lines)
and DeltaTableRefreshAndPinningConnectSuite (1964 lines) into focused
category traits following the Spark DSv2 test pattern from PRs 55536,
55462, 55463, 55540, 55571.

Classic module (spark/src/test/scala/.../delta/):
- DeltaTableRefreshTestBase: shared helpers (table setup, external commits)
- DeltaTempViewRefreshTests: Section [1] temp views (16 tests)
- DeltaRepeatedAccessRefreshTests: Section [2] repeated access (6 tests)
- DeltaJoinRefreshTests: Section [3] join queries (8 tests)
- DeltaDatasetPinningTests: Section [4] show/collect pinning (9 tests)
- DeltaCacheTableRefreshTests: Section [5] CACHE TABLE (9 tests)
- DeltaTableRefreshAndPinningSuite: slim runner (6 concrete suites)

Connect module (spark-connect/client/src/test/scala/.../tables/):
- DeltaTableRefreshConnectTestBase: shared helpers (filesystem writes)
- DeltaTempViewRefreshConnectTests: Section [1] (15 tests)
- DeltaRepeatedAccessRefreshConnectTests: Section [2] (6 tests)
- DeltaJoinRefreshConnectTests: Section [3] (32 tests, 4 sub-groups)
- DeltaDatasetReanalysisConnectTests: Section [4] re-analysis (9 tests)
- DeltaCacheTableRefreshConnectTests: Section [5] (7 tests)
- DeltaTableRefreshAndPinningConnectSuite: slim runner (4 concrete suites)

Pure refactoring with no behavioral changes. Test counts preserved:
classic 48 tests x 6 suites, connect 69 tests x 4 suites.
